+++ src/app/Http/Controllers/Api/V1/Addresses/DistrictController.php
@@ -8,11 +8,15 @@ use App\Http\Requests\V1\Addresses\UpdateDistrictRequest;
 use App\Http\Resources\V1\Addresses\DistrictCollection;
 use App\Http\Resources\V1\Addresses\DistrictResource;
 use App\Models\V1\Addresses\District;
-use Illuminate\Http\Request;
 use Spatie\QueryBuilder\QueryBuilder;
 
 class DistrictController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
         $districts = QueryBuilder::for(District::class)
@@ -27,7 +31,9 @@ class DistrictController extends Controller
 
     public function store(StoreDistrictRequest $request)
     {
+        $this->authorize('create', District::class);
         District::create($request->validated());
+
         return response(['message' => 'District successfully create'], 201);
     }
 
@@ -38,12 +44,16 @@ class DistrictController extends Controller
 
     public function update(UpdateDistrictRequest $request, District $district)
     {
+        $this->authorize('update', $district);
         $district->update($request->validated());
+        
         return response(['message' => 'District successfully updated'], 200);
     }
 
     public function destroy(District $district)
     {
+        $this->authorize('delete', $district);
+
         return $district->delete()
             ? response()->noContent()
             : response()->json(['message' => 'District not deleted'], 400);
