+++ src/app/Http/Controllers/Api/V1/Addresses/RegionController.php
@@ -3,18 +3,20 @@
 namespace App\Http\Controllers\Api\V1\Addresses;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\V1\Addresses\StoreCountryRequest;
 use App\Http\Requests\V1\Addresses\StoreRegionRequest;
 use App\Http\Requests\V1\Addresses\UpdateRegionRequest;
 use App\Http\Resources\V1\Addresses\RegionCollection;
 use App\Http\Resources\V1\Addresses\RegionResource;
 use App\Models\V1\Addresses\Region;
-use Exception;
-use Illuminate\Http\Request;
 use Spatie\QueryBuilder\QueryBuilder;
 
 class RegionController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
         $regions = QueryBuilder::for(Region::class)
@@ -29,7 +31,9 @@ class RegionController extends Controller
 
     public function store(StoreRegionRequest $request)
     {
+        $this->authorize('create', Region::class);
         Region::create($request->validated());
+        
         return response()->json(['message' => 'Region successfully created'], 201);
     }
 
@@ -40,12 +44,16 @@ class RegionController extends Controller
 
     public function update(UpdateRegionRequest $request, Region $region)
     {
+        $this->authorize('update', $region);
         $region->update($request->validated());
+
         return response()->json(['message' => 'Region seccessfully updated'], 200);
     }
 
     public function destroy(Region $region)
     {
+        $this->authorize('delete', $region);
+
         return $region->delete()
             ? response()->noContent()
             : response()->json(['message' => 'Region not deleted'], 400);
