+++ src/app/Http/Controllers/Api/V1/Books/BookController.php
@@ -17,6 +17,7 @@ class BookController extends Controller
         private BookRepositoryInterface $repository
     )
     {
+        $this->middleware('auth:sanctum', ['except' => ['index', 'show']]);
     }
 
     public function index()
@@ -26,6 +27,8 @@ class BookController extends Controller
 
     public function store(StoreBookRequest $request)
     {
+        $this->authorize('create', Book::class);
+
         return $this->repository->store($request->validated())
             ? response()->json(['message' => 'Book successfully created'], 201)
             : response()->json(['message' => 'Book not created'], 400);
@@ -38,6 +41,8 @@ class BookController extends Controller
 
     public function update(UpdateBookRequest $request, Book $book)
     {
+        $this->authorize('update', $book);
+
         return $this->repository->update($book, $request->validated())
             ? response()->json(['message' => 'Book successfully updated'], 200)
             : response()->json(['message' => 'Book not updated'], 400);
@@ -45,6 +50,8 @@ class BookController extends Controller
 
     public function destroy(Book $book)
     {
+        $this->authorize('delete', $book);
+
         return $this->repository->destroy($book)
             ? response()->noContent()
             : response()->json(['message' => 'Book not deleted'], 400);
@@ -52,6 +59,8 @@ class BookController extends Controller
 
     public function deleteFormat(Book $book, string $format)
     {
+        $this->authorize('deleteFormat', $book);
+
         $format = ucfirst($format);
 
         if (!BookFormat::tryFrom($format)) {
@@ -61,5 +70,15 @@ class BookController extends Controller
         return $book->deleteFormat(BookFormat::from($format))
             ? response()->noContent()
             : response()->json(['message' => 'Format not deleted'], 400);
+    }
+
+    public function buy(Book $book, string $format)
+    {
+        // create token or create token in order method
+    }
+
+    public function download(Book $book, string $format)
+    {
+        // check token and download the book
     }
 }
