+++ src/app/Http/Controllers/Api/V1/Books/PublisherController.php
@@ -13,12 +13,17 @@ use Spatie\QueryBuilder\QueryBuilder;
 
 class PublisherController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum', ['except' => ['index', 'show']]);
+    }
+
     public function index()
     {
         $publishers = QueryBuilder::for(Publisher::class)
             ->allowedFilters([AllowedFilter::exact('id'), 'name'])
             ->allowedFields(['id', 'name'])
-            ->allowedSorts(['name'])
+            ->allowedSorts(['id', 'name'])
             ->get();
 
         return new PublisherCollection($publishers);
@@ -26,8 +31,10 @@ class PublisherController extends Controller
 
     public function store(StorePublisherRequest $request)
     {
+        $this->authorize('create', Publisher::class);
         Publisher::create($request->validated());
-        return response()->json(['message' => 'Publshier successfully created', 201]);
+
+        return response()->json(['message' => 'Publshier successfully created'], 201);
     }
 
     public function show(Publisher $publisher)
@@ -37,12 +44,16 @@ class PublisherController extends Controller
 
     public function update(UpdatePublisherRequest $request, Publisher $publisher)
     {
+        $this->authorize('update', $publisher);
         $publisher->update($request->validated());
+        
         return response()->json(['message' => 'Publisher successfully updated'], 200);
     }
 
     public function destroy(Publisher $publisher)
     {
+        $this->authorize('delete', $publisher);
+
         return $publisher->delete() 
             ? response()->noContent()
             : response()->json(['message' => "Publisher not deleted"], 500);
